Automate disk type selection with NVMe preference
sedutil uses the serial number to hash the password for NVMe disk types
and the model number for SCSI disk types. Tri-mode NVMe devices are
exposed as SCSI devices, making them incompatible on non-tri-mode HBAs.
This patch automates disk type selection by first checking if the disk
is compatible with the NVMe IOCTL interface using the NVME_IOCTL_ID
command. If it's not, it falls back to SCSI.